Completely deleted the tavern NPC and moved the shop to the town's magister
